requirements.txt
aiohttp==1.3.3
PyYAML==3.12
